changge TwisterInput to only deal with 1-127 velocity values (not floats)

+++ GroupCtl/src/main/java/dev/tradcode/groupctl/events/EncoderTurned.java
@@ -1,3 +1,3 @@
 package dev.tradcode.groupctl.events;
 
-public record EncoderTurned(int n, float val) implements Event { }
+public record EncoderTurned(int n, int val) implements Event { }
